refactor for new dpf - wip

+++ plugins/Ninjas2/Widgets/NanoWheel.hpp
@@ -57,7 +57,7 @@ class NanoWheel : public WolfWidget
     Callback *fCallback;
 
     bool fLeftMouseDown;
-    Point<int> fLeftMouseDownLocation;
+    Point<double> fLeftMouseDownLocation;
 
     bool fIsHovered;
